requirements.txt
Django==3.2.4
psycopg2-binary==2.9.1
django-ckeditor==6.1.0
django-braces==1.14.0
django-q==1.3.8
django-simple-history==3.0.0
django-getenv>=1.3.1
django-memoize==2.3.1
eea.frame==0.5.5
redis==3.5.3
raven==6.10.0
